flask-backend/requirements.txt
pandas
flask
openai
python-dotenv
langchain
langchain_experimental
langchain_openai
openpyxl
tabulate
flask-cors
